+++ index.html
@@ -53,6 +53,7 @@
   </div>
 <div class="main--movie">
   <div id="movieFeed" class="main--movie--item">
+  <!-- <button class="main--movie--item--action--btn" id="moreBtn">More</button> -->
   </div>
 </div>
  </div>
