+++ js/app.js
@@ -1,14 +1,15 @@
-import renderMovies from "./renderMovies";
+import renderMovies from "./renderMovies.js";
+import makeMoreButton from "./moreButton.js"
 
 const movieFeed = document.getElementById('movieFeed');
 
 // Search section
-const searchEl = document.getElementById('searchGroup');
-const inputEl = document.getElementById('search-input')
-const buttonEl = document.getElementById('searchBtn')
+const inputEl = document.getElementById('search-input');
+const buttonEl = document.getElementById('searchBtn');
+
 
 // 영화 검색 결과 불러오기
-async function getMovies(title, year='', page = 2) {
+async function getMovies(title, year='', page = 1) {
     const s = `&s=${title}`
     const y = `&y=${year}`
     const p = `&page=${page}`
@@ -17,51 +18,9 @@ async function getMovies(title, year='', page = 2) {
     console.log(movies);
     console.log(totalResults);
     renderMovies(movies);
-}
-
-// 검색 결과 출력
-// function renderMovies(movies) {
-//       for(let movie of movies){
-//         const movieEl = document.createElement('div');
-//         const imgDivEl = document.createElement('div');
-//         const imgEl = document.createElement('img');
-//         const infoDivEl = document.createElement('div');
-//         const titleEl = document.createElement('h4');
-//         const yearEl = document.createElement('p');
-//         movieEl.className = 'main--movie--item--group';
-//         imgDivEl.className = 'main--movie--item--group--imgdiv';
-//         infoDivEl.className = 'main--movie--item--group--infodiv';
-//         titleEl.className = 'main--movie--item--group--infodiv--title';
-//         yearEl.className = 'main--movie--item--group--infodiv--year';
-//         imgEl.src = `${movie.Poster}`;
-//         titleEl.innerText = `${movie.Title}`;
-//         yearEl.innerText = `${movie.Year}`;
-//         imgDivEl.appendChild(imgEl);
-//         infoDivEl.appendChild(titleEl);
-//         infoDivEl.appendChild(yearEl);
-//         movieEl.appendChild(imgDivEl);
-//         movieEl.appendChild(infoDivEl);
-//         movieFeed.appendChild(movieEl);
-//         movieFeed.appendChild(actionDiv);
-//       };
-// }
-
-// 결과 더 가져오기 버튼
-// const actionDiv = document.createElement('div');
-// actionDiv.className = 'main--movie--item--action';
-// const moreBtn = document.createElement('button');
-// moreBtn.innerHTML = 'More';
-// moreBtn.className = 'main--movie--item--action--btn';
-// actionDiv.appendChild(moreBtn);
-// moreBtn.addEventListener('click', clickMoreBtn);
-
-// function clickMoreBtn() {
-//   // const movies = getMovies(page);
-//   // const pageUp = page+1;
-//   // getMovies(pageUp);
-//   // renderMovies(movies);
-//   console.log('page2')
-// }
+    makeMoreButton();
+    return movies;
+};
 
 // search 버튼 클릭 시 input의 value가 api {t}로 전송
 buttonEl.addEventListener('click', searchClick);
